refactor config module
